unblocked GameObjectTests clean up before pull request

+++ Project/Termgine/API/Scene.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Numerics;
 using System.Text;
 
 namespace Termgine {
@@ -89,10 +88,10 @@ namespace Termgine {
             if (gameObjectContent == null)return;
             var contentLines = gameObjectContent.Split('\n');
             for (var y = 0; y < contentLines.Length; y++) {
-                var globalPositionY = (int)(y + position.Y);
+                var globalPositionY = y + position.Y;
                 if (globalPositionY >= Height)continue;
                 for (var x = 0; x < contentLines[y].Length; x++) {
-                    var globalPositionX = (int)(x + position.X);
+                    var globalPositionX = x + position.X;
                     if (globalPositionX >= Width)break;
                     if (contentLines[y][x] != ' ' && contentLines[y][x] != '\0')
                         _content[globalPositionY][globalPositionX] = contentLines[y][x];
@@ -104,10 +103,10 @@ namespace Termgine {
             if (gameObjectColorMask == null)return;
             var contentLines = gameObjectColorMask.Split('\n');
             for (var y = 0; y < contentLines.Length; y++) {
-                var globalPositionY = (int)(y + position.Y);
+                var globalPositionY = y + position.Y;
                 if (globalPositionY >= Height)continue;
                 for (var x = 0; x < contentLines[y].Length; x++) {
-                    var globalPositionX = (int)(x + position.X);
+                    var globalPositionX = x + position.X;
                     if (globalPositionX >= Width)break;
                     if (contentLines[y][x] != ' ' && contentLines[y][x] != '\0')
                         _colorMask[globalPositionY][globalPositionX] = contentLines[y][x];
